pyproject.toml
[project]
name = "bookstore"
version = "0.1.0"
description = "Bookstore API"
authors = [
    {name = "Julio Silva",email = "silvajulio@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "djangorestframework (>=3.15.2,<4.0.0)",
    "django (>=5.1.6,<6.0.0)",
    "django-extensions (>=3.2.3,<4.0.0)",
    "django-debug-toolbar (>=5.0.1,<6.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "flake8 (>=7.1.1,<8.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
factory-boy = "^3.3.3"
django = "^5.1.6"

